Another attempt at pushing images...
Very odd - the 'tags' list is still '15.3, 15', not the values I
explicitly provide at the top of main.yml.

So for now there's two hacks. The first is explicitly creating a new
list, 'actual_versions', in vars/main.yml. (Just in case there's a
problem getting the parameters in to the 'push-images' playbook.)

The second is moving the contents of 'push-images' into main.