Added minimal TestContainer demo
Added minimal TestContainer demo (maven project).

This is a partial port of an existing project that captures software package information
on Debian-based systems, with the primary focus at the moment with showing how to use
TestContainers and FlywayDB scripts to launch and configure a test database.

Testing code using database server extensions is simple - we simply start with a docker
image that has already installed the required package, e.g., `postgresql-17-pljava`,
and then put `CREATE EXTENSION...` in your flyway script.

I'm using jOOQ, not JPA, since I think there may be a way to have the same java
classes used in both the database and the application if we have the proper
jOOQ binding.